ml_pipelines/requirements.txt
# ML Pipeline Dependencies
# 
# Core dependencies for the ML pipeline with full Vertex AI and TensorBoard support

# Google Cloud & Vertex AI
google-cloud-bigquery>=3.0.0
google-cloud-bigquery-storage>=2.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.0.0

# Kubeflow Pipelines (KFP)
kfp>=2.0.0

# TensorFlow & TensorBoard
tensorflow>=2.14.0
tensorboard>=2.14.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=10.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Database
db-dtypes>=1.0.0

# Utilities
tqdm>=4.65.0
